refactor(app): lift effect clusters out of App.tsx into hooks (#316 phase 2)
Follows the useStartupData pattern already in hooks/ — state and effects
lifted out, values and setters returned. Pure moves; dependency arrays are
unchanged.

- useServiceWorkerUpdate(screen): useRegisterSW, swRegRef, updateDismissed,
  the controllerchange reload listener, and the title-screen update poll.
  Also exposes checkForUpdates(), which absorbs the inline
  SettingsScreen onCheckForUpdates body, so swRegRef leaves App entirely.
  Prompt-mode behaviour (no skipWaiting, no auto-apply) is untouched.
- useTabVisibility({ screenRef, currentLocationKeyRef }): isTabHidden plus
  the "returned from background" Rollbar breadcrumb. Reads screen/location
  through refs so the listener stays bound once at mount, as before.
- useScreenGuards({...}): the data-dependent screen bail-out chain.
- useTownAccess(isAdmin): enabledTownIds fetch, previewAsPlayer,
  bypassTownAccess, and the restrictedTownNodeIds fog computation.
- useBattleTelemetry({...}): unit-death tracking, kill-driven achievement
  and quest progress, flawless flag, both auto-slowdown rules, and the
  obituary/tired-game/time-capsule secrets.

Refs used only within a moved cluster (hiddenAtRef, prevCommanderHpRef,
prevPendingCastKeyRef, timeCapsuleCheckedRef) now live inside their hook.
Refs App resets from many battle-start paths (prevPlayerUnitsRef,
prevOpponentUnitsRef, battleFlawlessRef) stay owned by App and are passed in.

App.tsx: 4,030 -> 3,734 lines. Build green, chunk boundaries unchanged,
2107 tests passing.
